feat: extract accessible risk visualisations
